Move example iterators to a new module
