+++ java/compaction/compaction-job-execution/src/test/java/sleeper/compaction/job/execution/JavaCompactionRunnerIteratorIT.java
@@ -21,12 +21,12 @@ import sleeper.compaction.core.job.CompactionJob;
 import sleeper.compaction.job.execution.testutils.CompactionRunnerTestBase;
 import sleeper.compaction.job.execution.testutils.CompactionRunnerTestData;
 import sleeper.compaction.job.execution.testutils.CompactionRunnerTestUtils;
-import sleeper.core.iterator.impl.AgeOffIterator;
 import sleeper.core.partition.PartitionsBuilder;
 import sleeper.core.record.Record;
 import sleeper.core.schema.Schema;
 import sleeper.core.statestore.FileReference;
 import sleeper.core.tracker.job.run.RecordsProcessed;
+import sleeper.example.iterator.AgeOffIterator;
 import sleeper.sketches.testutils.SketchesDeciles;
 
 import java.util.List;
