+++ java/example-iterators/src/test/java/sleeper/example/iterator/AgeOffIteratorTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.core.iterator.impl;
+package sleeper.example.iterator;
 
 import org.junit.jupiter.api.Test;
