+++ java/query/query-runner/pom.xml
@@ -67,6 +67,12 @@
         </dependency>
         <dependency>
             <groupId>sleeper</groupId>
+            <artifactId>example-iterators</artifactId>
+            <version>${project.parent.version}</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>sleeper</groupId>
             <artifactId>dynamodb-test</artifactId>
             <version>${project.parent.version}</version>
             <scope>test</scope>
